+++ pwnscripts/fsb/__init__.py
@@ -1,3 +1,5 @@
 '''Everything related to extending pwntools' fmtstr module'''
 import pwnscripts.fsb.find_offset
 import pwnscripts.fsb.leak
+PRINTF_MIN = 5
+PRINTF_MAX = 35